A couple minor fixes (#7299)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request

- If you start a `me` message with a comma or an apostrophe, there will
no longer be a space between your name and the message. I.e. "**John
Trasen** , being hungry, eats." -> "**John Trasen**, being hungry,
eats."
- Fixes reversed logic with dropping things with a broken hand. Your
character will now correctly cry in pain when trying to do this.
- Also potentially fixes reversed logic with collapsing with a broken
leg and screaming in pain.
- In both of the above cases, it will deal halloss (pain) damage, since
using a broken bone tends to hurt a lot.

## Why It's Good For The Game

Small fix and improvement.

## Changelog

<!-- If your PR modifies aspects of the game that can be concretely
observed by players or admins you should add a changelog. If your change
does NOT meet this description, remove this section. Be sure to properly
mark your PRs to prevent unnecessary GBP loss. You can read up on GBP
and it's effects on PRs in the tgstation guides for contributors. Please
note that maintainers freely reserve the right to remove and add tags
should they deem it appropriate. You can attempt to finagle the system
all you want, but it's best to shoot for clear communication right off
the bat. -->

:cl:
fix: Fixed erroneous space when starting a me message with a comma or
apostrophe.
fix: Your character will now cry in pain when trying to hold something
with a broken hand.
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->