Caparisons (#7296)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request

Adds a single caparison to cover quadruped lower bodies. It's only
codersprites reusing the poncho that's included for every taur icon
sheet, cutting away what's not needed and painting over it in greyscale.
Wip as only some are done, and I'm opening it early incase someone may
have a better icon for at least the item, which is just a poncho with
the same greyscale.
It also pushes the tail layer one down, so any taurized clothing doesn't
get on the same layer as their saddlebags, which in turn has the last
worn of the two be rendered over the other.
Taur bodies supported:
- [x] feline 
- [x] cow
- [x] drake
- [x] deer
- [x] wolf
- [x] otie
- [x] lizard
- [x] horse 

## Why It's Good For The Game

There currently are no clothes that would end up covering a taur body
that would work for most, and it's something not as specific as hardsuit
otherwise available.

## Changelog

<!-- If your PR modifies aspects of the game that can be concretely
observed by players or admins you should add a changelog. If your change
does NOT meet this description, remove this section. Be sure to properly
mark your PRs to prevent unnecessary GBP loss. You can read up on GBP
and it's effects on PRs in the tgstation guides for contributors. Please
note that maintainers freely reserve the right to remove and add tags
should they deem it appropriate. You can attempt to finagle the system
all you want, but it's best to shoot for clear communication right off
the bat. -->

:cl:
add: Caparisons, a body covering for taurs
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

---------

Co-authored-by: MarsMond <no@email>