Three Freaks and Fixes (#7313)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## Stuff for Halloween This Year
For Halloween this year, I have three new mobs here to provide
additional challenges to the crew.

First up the Cyclops: The Blood Cult's very own mech equivalent,
equipped with deadly penetrating eye beams. It can launch a telegraphed
death laser to destroy cover and melt its enemies in seconds. Designed
to actually be a threat to people with ranged weapons, it will hopefully
not disappoint.

Secondly the Alphawraith: Similar to the Gammawraith, this 'living'
paracasual representation of an alpha particle delivers power beam
attacks with low AP in contrast to the Gammawraith's weak attacks with
high AP.

Finally, the Wendigo: A fast, melee death blender that has a leaping and
cloaking variant. It is highly resistant to brute damage, so use beams
to melt through its health.

All of these are still being balanced so some of their abilities and
strengths may change depending on feedback and their in-game
performance.

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Other Miscellaneous
The Airlock Mimic's sprite now resembles the new airlock art style. The
Gammawraith has a new 4 direction sprite and a buffed firerate. A
Lythios POI meant for caves is now indoors, and Lythios cave POI budget
has also been increased.

<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->

## Changelog

<!-- If your PR modifies aspects of the game that can be concretely
observed by players or admins you should add a changelog. If your change
does NOT meet this description, remove this section. Be sure to properly
mark your PRs to prevent unnecessary GBP loss. You can read up on GBP
and it's effects on PRs in the tgstation guides for contributors. Please
note that maintainers freely reserve the right to remove and add tags
should they deem it appropriate. You can attempt to finagle the system
all you want, but it's best to shoot for clear communication right off
the bat. -->

:cl:
add: 3 New Paracasual mobs
tweak: Sprites for Gammawraith and Door Mimic, Lythios POI spawn budget.
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->