Remake Atlas FD Office (#7327)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request

Changes the FD's office and the area outside it to have a small
lounge/cafe area and gives the FD's office some windows.

This also fixes a nightshift setting I missed in medical's surgery
hallway + fixes the disconnected APC I added to the library hallway the
other day.

## Why It's Good For The Game

Not having any windows and being sequestered away isn't really befitting
for the FD's office. This helps make it more inviting and gives
"natural" light, plus gives a place for command staff to hang out other
than the bridge/meeting room.

The other two are minor fixes.
## Changelog

<!-- If your PR modifies aspects of the game that can be concretely
observed by players or admins you should add a changelog. If your change
does NOT meet this description, remove this section. Please note that
maintainers freely reserve the right to remove and add tags should they
deem it appropriate. You can attempt to finagle the system all you want,
but it's best to shoot for clear communication right off the bat. -->

:cl:
tweak: Changed around the FD's office on the Atlas map, and added a new
lounge area outside of it.
fix: Fixed the surgery hallway not obeying the nightshift setting.
fix: Fixed the APC in the library hallway on Atlas not being connected
properly.
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->