Updates nightshift colors (#7347)
- Changes overall color back to warm colors
- Moves color, power, and range to a proper define
- Small bulbs and fairy lights now respect the nightshift

<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request

Updates the nightshift colors to be warm again (pending the conclusion
of a vote in devlog). Creates more defines for nightshift color
settings, and allows small lightbulbs and fairy lights (which are the
main lightsources alongside tubes) to respect the nightshift as well.

## Why It's Good For The Game

Letting small lights and fairy lights be affected by the nightshift is
arguably good, especially on maps like the Strelka where they make up
the majority of light sources. Making the colors warmer can help make
the nightshift more cozy, though it is still up to a vote.

## Changelog

<!-- If your PR modifies aspects of the game that can be concretely
observed by players or admins you should add a changelog. If your change
does NOT meet this description, remove this section. Please note that
maintainers freely reserve the right to remove and add tags should they
deem it appropriate. You can attempt to finagle the system all you want,
but it's best to shoot for clear communication right off the bat. -->

:cl:
tweak: Small lightbulbs and fairy lights now respect the nightshift.
tweak: The nightshift color is now warm instead of cold.
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->