Big Mech Time/Quasimodo Addition (WIP) (#7334)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request
This prepares the grounds for 3x3 mechs to be added into the game,
starting off with the first trial of the Quasimodo. The Quasimodo is
designed to carry a single large, anti tank weapon, with minimal room
for other equipment or weapons. The idea is to make it large and tanky,
but at the cost of it not being able to be repaired unless attached to a
mech platform to hold it(To be Added). The weapons planned to be
attached to it is an explosive, Anti Tank/Anti Building Cannon, with
more like it along the way.
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
I'll be honest, this was mainly for fun. I remember that people have
been wanting larger mechs for a while, so I wanted to try my hand at it.
At the very least, if my mech doesn't get approved for crew creation,
it'll still lay the grounds for other coders to create larger mechs for
events/crew usage.
<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->

## Changelog

- ✓ Adds the Quasimodo!

- ✓ Adds the HAG-30(AC-20) 

- ✓ Adds a new mech weapon class! (Heavy Arms)

- (WIP) Adds the mech platform to create and repair the new 3x3 mechs
with!

- ? Doesn't send the server into hell over the balance of 3x3 mechs!
(Optional)



https://github.com/user-attachments/assets/2a811bc3-0076-4561-ab22-1049bff95d61



https://github.com/user-attachments/assets/8872563e-6ed3-43e7-924f-e6b902556171



https://github.com/user-attachments/assets/9f6513c5-cc56-4249-8a16-284fbf2442c7


<!-- If your PR modifies aspects of the game that can be concretely
observed by players or admins you should add a changelog. If your change
does NOT meet this description, remove this section. Please note that
maintainers freely reserve the right to remove and add tags should they
deem it appropriate. You can attempt to finagle the system all you want,
but it's best to shoot for clear communication right off the bat. -->

:cl: Kade
add: Adds the new 3x3 mech class and the Quasimodo to go along with it.
Also includes a weapon class (Heavy) to be used along side the 3x3
mechs.
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

---------

Co-authored-by: silicons <2003111+silicons@users.noreply.github.com>