Rebalances traveler voidsuits and removes an adminspawn weapon (#7390)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request

This PR adds a variety of voidsuit for the travelers off-maps that is
not straight up Syndicate gear with the associated stats; it now just
*looks* like the Syndicate gear while being more reasonably balanced.

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game

Will stop random offmaps from having access to antag-level gear. 

<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->

## Changelog

<!-- If your PR modifies aspects of the game that can be concretely
observed by players or admins you should add a changelog. If your change
does NOT meet this description, remove this section. Please note that
maintainers freely reserve the right to remove and add tags should they
deem it appropriate. You can attempt to finagle the system all you want,
but it's best to shoot for clear communication right off the bat. -->

:cl:
add: Traveler.dm to voidsuits lists, with downgraded syndie suits and
helmets for travelers to use
tweak: upped the prices on the suits in the vendor on Nebula to more
reasonable levels
add: added a new armor class for traveler suits
tweak: upped the armor values for Exploration suits
tweak: rebalanced the RHR on Lythious 43a to a more reasonable weapon
tweak: fixed airlocks on Vevalia shuttle and syndie suit displays to the
traveler variant
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->