remove dead branch in blood.dm
uses a datum[varname] which is no longer supported, and also unchecked. data.species would have revealed that the datum data is does not have the species keyword 9 months ago.

Blood mixtures dont have a species keyword so this branch is not possible to check afaik and pointless to have.