adds a bandaid fix for micro taur scaling (#7405)
## About The Pull Request
adds a bandaid fix for micro taur scaling

the issue is specifically if you are scaling below 100% with taur parts
that have icon sidedness
there's only one taur part with this and it's a lamia tail that i use

there's only so much i can do because it seems to be a byond issue

## Why It's Good For The Game
fixes a bug that only affects me

## Changelog

<!-- If your PR modifies aspects of the game that can be concretely
observed by players or admins you should add a changelog. If your change
does NOT meet this description, remove this section. Please note that
maintainers freely reserve the right to remove and add tags should they
deem it appropriate. You can attempt to finagle the system all you want,
but it's best to shoot for clear communication right off the bat. -->

:cl:
fix: adds a bandaid fix for micro taur scaling
/:cl: