Nerf Hardsuit Sprinter Module by 50% (#7428)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request

speed demons are verboten

## Why It's Good For The Game

if your character has the ability to become the fastest character in the
game through un-encumberance stacking then that is not the fault of the
characters but the fault of the game

this fixes an issue with the game wherein players can exploit this
caveat. nevermind the fact that a hardsuit shouldnt let you move that
fast. your legs would break. or joints. both.

speed variable cut by half. it lowers the hardsuit roughly down to
teshari speed rather than above it. likewise power consumption was
increased by 150% given the strength of movement

if you ask me they shouldnt even get to be teshari/tajaran speed but
this is an amicable middle point
 
## Changelog

<!-- If your PR modifies aspects of the game that can be concretely
observed by players or admins you should add a changelog. If your change
does NOT meet this description, remove this section. Please note that
maintainers freely reserve the right to remove and add tags should they
deem it appropriate. You can attempt to finagle the system all you want,
but it's best to shoot for clear communication right off the bat. -->

:cl:
balance: nerfed sprinter module
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->